+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 2023-12-03
+## [15.3.0] 2023-12-08
 ### Removed
 - AI folder, migrated to nodebackend.
 - Removed exclusion of test folder from tsconfig.
@@ -17,9 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Excluded tests folder from build.
 - Renamed folder 'tests' to 'test'.
 - Set `testMatch` in package.json to avoid running tests twice.
+- Migration of inputValidation to TS
 
 ### Added
 - Extended test case with Puppeteer to click a switch.
+- Extended test case with Puppeteer to toggle switch on and off
 
 ## [v15.2.1] 2023-12-02
 ### Changed